review: re-add lodash for non-trivial helpers, dedupe scryfall fetch client, mark adapter shim as temporary
Addresses PR #178 review feedback:

- util.mjs: keep round/clamp native (genuine Math.round/min/max
  one-liners), re-add lodash for orderBy/mean/pick/omit/once instead of
  hand-rolling reimplementations of them. @dills1220/nedb already pulls
  lodash in transitively, so this costs nothing extra in install size --
  lodash usage is now consolidated to this one file instead of scattered.
- image-processing/image-processor.mjs: use util.mjs's once (lodash)
  instead of a hand-rolled once() helper.
- scryfall-api/http-client.mjs (new): shared native-fetch request()
  wrapper + REQUEST_HEADERS, replacing the identical copy that used to
  live in both searchName.mjs and getCardName.mjs.
- storage/adapters/create-callback-adapter.mjs: marked explicitly as a
  temporary migration shim with the real fix path documented (make
  db-local/rds store modules natively promise-based so nothing needs to
  promisify a callback API downstream). Follow-up tracked separately.
