+++ package.json
@@ -75,6 +75,7 @@
         "image-size": "^1.2.1",
         "jimp": "0.22.12",
         "joi": "^18.2.3",
+        "lodash": "^4.18.1",
         "mysql2": "^3.16.0",
         "string-similarity": "^4.0.4",
         "tesseract.js": "^3.0.3"
