+++ src/util.mjs
@@ -1,3 +1,4 @@
+import _ from "lodash";
 import txtUtil from "clean-text-utils";
 import { createRequire } from "node:module";
 import logger from "./logger/log.mjs";
@@ -24,54 +25,25 @@ function requireOrFalse(modulePath) {
     }
 }
 
-// Small native replacements for the handful of lodash helpers used repeatedly across the
-// codebase -- kept here once instead of pulling in lodash for a handful of one-line functions.
-
+// round/clamp are genuine one-liners over Math.round/min/max, worth keeping native. Multi-key
+// sort, mean, pick, and omit are not -- reimplementing (and re-testing) lodash's versions of
+// those isn't worth it, and @dills1220/nedb already pulls lodash in transitively, so using it
+// here for the non-trivial pieces costs nothing extra in install size.
 function round(value, precision = 0) {
-    const factor = 10 ** precision;
-    return Math.round((value + Number.EPSILON) * factor) / factor;
+    return Math.round((value + Number.EPSILON) * 10 ** precision) / 10 ** precision;
 }
 
 function clamp(value, lower, upper) {
     return Math.min(Math.max(value, lower), upper);
 }
 
-function mean(numbers = []) {
-    return numbers.reduce((sum, n) => sum + n, 0) / numbers.length;
-}
+const mean = _.mean;
+const pick = _.pick;
+const omit = _.omit;
+const orderBy = _.orderBy;
+const once = _.once;
 
-function pick(object, keys = []) {
-    return keys.reduce((result, key) => {
-        if (object != null && Object.prototype.hasOwnProperty.call(object, key)) {
-            result[key] = object[key];
-        }
-        return result;
-    }, {});
-}
-
-function omit(object, keys = []) {
-    const excluded = new Set(keys);
-    return Object.fromEntries(Object.entries(object).filter(([key]) => !excluded.has(key)));
-}
-
-// Multi-key sort, e.g. orderBy(items, ["score", "name"], ["desc", "asc"]).
-function orderBy(collection, keySelectors = [], directions = []) {
-    const selectors = keySelectors.map((selector) =>
-        typeof selector === "function" ? selector : (item) => item[selector]
-    );
-    return [...collection].sort((left, right) => {
-        for (let index = 0; index < selectors.length; index += 1) {
-            const direction = directions[index] === "desc" ? -1 : 1;
-            const leftValue = selectors[index](left);
-            const rightValue = selectors[index](right);
-            if (leftValue < rightValue) return -direction;
-            if (leftValue > rightValue) return direction;
-        }
-        return 0;
-    });
-}
-
-export { cleanString, requireOrFalse, round, clamp, mean, pick, omit, orderBy };
+export { cleanString, requireOrFalse, round, clamp, mean, pick, omit, orderBy, once };
 
 export default {
     cleanString,
@@ -81,5 +53,6 @@ export default {
     mean,
     pick,
     omit,
-    orderBy
+    orderBy,
+    once
 };
